Azure Functions/GELF-Receiver-Prod/requirements.txt
azure-functions
python-dateutil
requests 